🚇🗑️ Remove binder link and integrations (#300)
This PR removes the [binder](https://mybinder.org) and the integration to add a PR comment with a binder link, as well as the workflow that triggers a build on binder when a commit is pushed to main.
The reason to remove binder is that it hasn't been working properly for many months (i.e. timeout when opening) and checking the link in the docs recently also started failing.