📚🩹 Use fixed pydata-sphinx-theme from fork (#334)
Today I found that our docs are broken due to a bug when building PDFs.

This PR changes the [pydata-sphinx-theme](https://github.com/pydata/pydata-sphinx-theme) from the published upstream project to a branch on my for including the following fixes:

- https://github.com/pydata/pydata-sphinx-theme/pull/2097
- https://github.com/pydata/pydata-sphinx-theme/pull/2076

I also went ahead and removed the workaround on our side since it isn't needed with the fix on my fork in place.

### Change summary

- [📚🩹 Use fixed pydata-sphinx-theme from fork](https://github.com/glotaran/pyglotaran-extras/commit/5f3bedfba3a9e765d857a9313e9f14cc25d7ba60)
- [🚇🩹 Setup docs env same as on RTD](https://github.com/glotaran/pyglotaran-extras/pull/334/commits/365899d39886ef2e8ae0dfdf94521e07ecf63a26)